feat - moved logic to series pass controler
