+++ src/Tickets/Flexible_Tickets/Base.php
@@ -25,7 +25,6 @@ use TEC\Tickets\Flexible_Tickets\Series_Passes\Reports;
 use Tribe__Tickets__Ticket_Object as Ticket_Object;
 use TEC\Tickets\Admin\Upsell as Ticket_Upsell;
 use TEC\Events_Pro\Custom_Tables\V1\Series\Provider as Series_Provider;
-use Tribe__Tickets__Tickets_View;
 use WP_Post;
 
 /**
@@ -182,8 +181,6 @@ class Base extends Controller {
 			$this->container->get( Series_Provider::class ),
 			'filter_remove_series_post_type'
 		] );
-
-		add_filter( 'tec_tickets_my_tickets_link_ticket_count_by_type', [ $this, 'filter_my_tickets_link_data' ], 10, 3 );
 	}
 
 	/**
@@ -282,8 +279,6 @@ class Base extends Controller {
 			$this->container->get( Series_Provider::class ),
 			'filter_remove_series_post_type'
 		] );
-
-		remove_filter( 'tec_tickets_my_tickets_link_ticket_count_by_type', [ $this, 'filter_my_tickets_link_data' ], 10, 3 );
 	}
 
 	/**
@@ -661,65 +656,4 @@ class Base extends Controller {
 		return tribe_format_date( $last_event->start_date, $display_time, $date_format );
 	}
 
-	/**
-	 * Filters the data for the "My Tickets" link.
-	 *
-	 * @since TBD
-	 *
-	 * @param array<string, mixed> $data The data for the "My Tickets" link.
-	 * @param int $event_id              The event ID.
-	 * @param int $user_id               The user ID.
-	 *
-	 * @return array<string, array> The updated data.
-	 */
-	public function filter_my_tickets_link_data( array $data, int $event_id, int $user_id ): array {
-		$post_type = get_post_type( $event_id );
-
-		// Only filter Events and Series, skip other post types.
-		if ( $post_type !== TEC::POSTTYPE && $post_type !== Series_Post_Type::POSTTYPE ) {
-			return $data;
-		}
-
-		// If we are on series page then replace ticket data with series counts.
-		if ( $post_type === Series_Post_Type::POSTTYPE ) {
-			$data['series'] = [
-				'count'    => $data['ticket']['count'],
-				'singular' => __( 'Pass', 'event-tickets' ),
-				'plural'   => __( 'Passes', 'event-tickets' ),
-			];
-
-			// Remove the ticket data.
-			$data['ticket']['count'] = 0;
-
-			return $data;
-		}
-
-		// Process series pass count for single event.
-		$series = tec_series()->where( 'event_post_id', $event_id )->first_id();
-
-		if ( empty( $series ) ) {
-			// Not part of a Series, bail.
-			return $data;
-		}
-
-		// Get the tickets purchased by this user and for this series.
-		$view = Tribe__Tickets__Tickets_View::instance();
-		$series_pass_count = $view->count_ticket_attendees( $series, $user_id );
-
-		if ( empty( $series_pass_count ) ) {
-			return $data;
-		}
-
-		$data['series'] = [
-			'count'    => $series_pass_count,
-			'singular' => __( 'Pass', 'event-tickets' ),
-			'plural'   => __( 'Passes', 'event-tickets' ),
-		];
-
-		if ( $data['ticket']['count'] > 0 ) {
-			$data['ticket']['count'] -= $series_pass_count;
-		}
-
-		return $data;
-	}
 }
